src/requirements.txt
asgiref==3.11.1
crispy-bootstrap4==2026.2
Django==5.2.12
django-crispy-forms==2.6
dotenv==0.9.9
mysql-connector==2.2.9
mysqlclient==2.2.8
pillow==12.1.1
python-dotenv==1.2.2
sqlparse==0.5.5
tzdata==2025.3
